aider: Created a new Go file `akeyless_auth.go` that contains a function `GetToken()` which checks for the existence of the `AKEYLESS_CLI_AUTHENTICATION_TOKEN_COMMAND` environment variable, executes the command, and returns the token or an error.
# Context:
USER: Please help me create a relatively simply golang library that can be utilized by CLI applications and a custom Terraform provider. The idea is to use an environment variable to shell out to the akeyless CLI to handle authentication and return a token that can be used to authenticate to the API. The idea is to not need to handle upgrading yet another tool to use all the auth methods supported by the CLI and instead just use the CLI. We need this library to support recognizing there is a "AKEYLESS_CLI_AUTHENTICATION_TOKEN_COMMAND" environment variable and use it to shell out to the CLI to get a token. The CLI will handle all the authentication methods supported by the CLI. The CLI will return a token that can be used to authenticate to the API. The CLI will be installed by the user separately and will be responsible for upgrading itself. Any errors should be returned by the library but the successful output of the command will return a token that can be used for authentication on all subsequent calls.
ASSISTANT: To create this library, we will need to create a new Go file. Let's call it `akeyless_auth.go`. This file will contain a function that checks for the existence of the `AKEYLESS_CLI_AUTHENTICATION_TOKEN_COMMAND` environment variable, executes the command, and returns the token or an error.

Here is the code for the new file:

```go
newfile/akeyless_auth.go
<<<<<<< ORIGINAL

=======
package akeyless

import (
	"os"
	"os/exec"
	"strings"
)

// GetToken shells out to the Akeyless CLI to handle authentication and return a token.
func GetToken() (string, error) {
	cmdStr := os.Getenv("AKEYLESS_CLI_AUTHENTICATION_TOKEN_COMMAND")
	if cmdStr == "" {
		return "", nil
	}

	cmdParts := strings.Fields(cmdStr)
	cmd := exec.Command(cmdParts[0], cmdParts[1:]...)

	output, err := cmd.Output()
	if err != nil {
		return "", err
	}

	token := strings.TrimSpace(string(output))
	return token, nil
}
>>>>>>> UPDATED
```

This function uses the `os/exec` package to execute the command specified in the `AKEYLESS_CLI_AUTHENTICATION_TOKEN_COMMAND` environment variable. The command's output is assumed to be the token. If the command fails for any reason, the error is returned. If the environment variable is not set, the function returns an empty string and no error.

Please let me know if you need any changes or if you have any questions.