aider: Return an error if the "AKEYLESS_CLI_AUTHENTICATION_TOKEN_COMMAND" is not set and reflect that error in the documentation.
# Context:
USER: return an error if the "AKEYLESS_CLI_AUTHENTICATION_TOKEN_COMMAND" is not set and reflect that error in the documentation
ASSISTANT: To implement this change, we need to modify two files: `akeyless/akeyless_auth.go` and `README.md`.

In `akeyless/akeyless_auth.go`, we will modify the `GetTokenFromAkeylessCommandLine` function to return an error if the `AKEYLESS_CLI_AUTHENTICATION_TOKEN_COMMAND` environment variable is not set.

In `README.md`, we will update the Usage section to reflect this change.

Here are the changes:

```go
akeyless/akeyless_auth.go
<<<<<<< ORIGINAL
func GetTokenFromAkeylessCommandLine() (string, error) {
	cmdStr := os.Getenv("AKEYLESS_CLI_AUTHENTICATION_TOKEN_COMMAND")
	if cmdStr == "" {
		return "", nil
	}
=======
func GetTokenFromAkeylessCommandLine() (string, error) {
	cmdStr := os.Getenv("AKEYLESS_CLI_AUTHENTICATION_TOKEN_COMMAND")
	if cmdStr == "" {
		return "", errors.New("AKEYLESS_CLI_AUTHENTICATION_TOKEN_COMMAND environment variable is not set")
	}
>>>>>>> UPDATED
```

```markdown
README.md
<<<<<<< ORIGINAL
To use Akeyless Sheller, run the `GetTokenFromAkeylessCommandLine` function. This function looks for the `AKEYLESS_CLI_AUTHENTICATION_TOKEN_COMMAND` environment variable to be set. The value of this variable should be something like:
=======
To use Akeyless Sheller, run the `GetTokenFromAkeylessCommandLine` function. This function requires the `AKEYLESS_CLI_AUTHENTICATION_TOKEN_COMMAND` environment variable to be set. If it is not set, the function will return an error. The value of this variable should be something like:
>>>>>>> UPDATED
```