aider: Updated the README.md file to include information about how the library will throw an error if the env variable is NOT set to a file that is executable.
# Context:
USER: we should update the docs too
ASSISTANT: Sure, I can help with that. Could you please provide more details about the changes you want to make to the documentation?
USER: we should update the docs on how the library will throw an error if the env variable is NOT set to a file that is executable
ASSISTANT: Understood. We will update the README.md file to include this information.

Here is the proposed change:

```markdown
README.md
<<<<<<< ORIGINAL
To use Akeyless Sheller, run the `GetTokenFromAkeylessCommandLine` function. This function requires the `AKEYLESS_CLI_AUTHENTICATION_TOKEN_COMMAND` environment variable to be set. If it is not set, the function will return an error. The value of this variable should be something like:
=======
To use Akeyless Sheller, run the `GetTokenFromAkeylessCommandLine` function. This function requires the `AKEYLESS_CLI_AUTHENTICATION_TOKEN_COMMAND` environment variable to be set. If it is not set, or if it is set to a path that does not point to an executable file, the function will return an error. The value of this variable should be something like:
>>>>>>> UPDATED
```

Please confirm if this change is acceptable.