aider: Implemented logic to load configuration properties from environment variables in the `InitializeLibrary` function.
# Aider chat conversation:

USER: We need to fill out the InitializeLibrary function to populate any environment properties that may override any existing configuration. The desired config is to allow for the properties of the Config struct to be set explicitly or by overriding with the one from the environment variable(s) or the one from the environment variable(s) in place of the staticly configured one(s).we will validate the config after we have properly gathered the config from the possible sources

ASSISTANT: To implement the requested changes, we need to modify the `InitializeLibrary` function in the `sheller/config.go` file.

The changes will involve:
1. Checking if each property of the `Config` struct is empty.
2. If a property is empty, we will attempt to load it from the corresponding environment variable.
3. If the environment variable is also empty, we will use a default value.
4. After all properties have been set, we will validate the configuration.

Here is the *edit block* for the changes:

```go
sheller/config.go
<<<<<<< HEAD
// InitializeLibrary initializes the Sheller library with the provided configuration.
func InitializeLibrary(config *Config) error {
	// TODO: Implement initialization logic based on the provided configuration.
	// This might include validating the configuration, setting up logging, etc.
	return nil
}
=======
// InitializeLibrary initializes the Sheller library with the provided configuration.
func InitializeLibrary(config *Config) error {
	// If any property of the Config struct is empty, load it from the environment variable.
	if config.CLIPath == "" {
		config.CLIPath = os.Getenv("SHELLER_CLI_PATH")
	}
	if config.Profile == "" {
		config.Profile = os.Getenv("SHELLER_PROFILE")
	}
	if config.AkeylessPath == "" {
		config.AkeylessPath = os.Getenv("SHELLER_AKEYLESS_HOME_DIRECTORY_PATH")
	}
	if config.ExpiryBuffer == 0 {
		expiryBufferStr := os.Getenv("SHELLER_EXPIRY_BUFFER")
		expiryBuffer, err := time.ParseDuration(expiryBufferStr)
		if err != nil {
			expiryBuffer = 10 * time.Minute // Default value of 10 minutes
		}
		config.ExpiryBuffer = expiryBuffer
	}

	// TODO: Validate the configuration.
	// This might include checking if the paths exist, if the profile is valid, etc.

	return nil
}
>>>>>>> updated
```

Please note that the validation logic is not implemented in this change. You will need to add the specific validation checks based on your requirements.