aider: Add a new `ValidateConfig` function that checks if `CLIPath` is set and if not, uses the `go-which` library to check if `akeyless` is in the system path, and update `InitializeLibrary` to call `ValidateConfig` after loading the configuration from environment variables.
# Aider chat conversation:

USER: we now need a new validate config function that gets called after we load env vars. to start with we need to make sure the cliPath is set, however if the path is NOT set BUT we can see the akeyless binary is in the system path then we know that is valid. we can use the library github.com/hairyhenderson/go-which to determine if akeyless is in the system path. we will add more validations after

ASSISTANT: To implement this, we need to modify the `sheller/config.go` file. We will add a new function `ValidateConfig` that will be called from the `InitializeLibrary` function after `LoadConfigFromEnv`. This function will check if `cliPath` is set and if not, it will use the `go-which` library to check if `akeyless` is in the system path.

First, we need to add the `go-which` library to our project. You can do this by running the following command in your terminal:

```bash
go get github.com/hairyhenderson/go-which
```

Then, we can proceed with the code changes.

```go
sheller/config.go
<<<<<<< HEAD
// InitializeLibrary initializes the Sheller library with the provided configuration.
func InitializeLibrary(config *Config) error {
	// Load configuration from environment variables
	LoadConfigFromEnv(config)

	// TODO: Validate the configuration.
	// This might include checking if the paths exist, if the profile is valid, etc.

	return nil
}
=======
import (
	"github.com/hairyhenderson/go-which"
	// other imports...
)

// ValidateConfig validates the provided configuration.
func ValidateConfig(config *Config) error {
	if config.CLIPath == "" {
		_, err := which.Which("akeyless")
		if err != nil {
			return fmt.Errorf("CLIPath is not set and akeyless is not in the system path")
		}
	}
	// TODO: Add more validations here.
	return nil
}

// InitializeLibrary initializes the Sheller library with the provided configuration.
func InitializeLibrary(config *Config) error {
	// Load configuration from environment variables
	LoadConfigFromEnv(config)

	// Validate the configuration.
	err := ValidateConfig(config)
	if err != nil {
		return err
	}

	return nil
}
>>>>>>> updated
```

This code adds the `ValidateConfig` function which checks if `CLIPath` is set in the configuration. If it's not set, it uses the `which.Which` function from the `go-which` library to check if `akeyless` is in the system path. If `akeyless` is not in the system path, it returns an error. The `InitializeLibrary` function is updated to call `ValidateConfig` after `LoadConfigFromEnv`. If `ValidateConfig` returns an error, `InitializeLibrary` returns that error.