aider: Added validation to check if the CLI path leads to an executable file.
# Aider chat conversation:

USER: now that we know that we have done everything we can to get the CLI path we need to validate that the path actually leads to an executable file and return an error if not

ASSISTANT: To validate that the CLI path leads to an executable file, we can use the `os` package's `Stat` function to get the file info and then check the `Mode` to see if it's an executable. We will add this validation in the `ValidateConfig` function in the `sheller/config.go` file.

Here is the change:

```go
sheller/config.go
<<<<<<< HEAD
// ValidateConfig validates the provided configuration.
func ValidateConfig(config *Config) error {
	if config.CLIPath == "" {
		akeylessFound := which.Which("akeyless")
		if akeylessFound == "" {
			return errors.New("the CLIPath is not set and akeyless is not in the system path")
		} else {
			config.CLIPath = akeylessFound
		}
	}
	// TODO: Add more validations here.
	return nil
}
=======
// ValidateConfig validates the provided configuration.
func ValidateConfig(config *Config) error {
	if config.CLIPath == "" {
		akeylessFound := which.Which("akeyless")
		if akeylessFound == "" {
			return errors.New("the CLIPath is not set and akeyless is not in the system path")
		} else {
			config.CLIPath = akeylessFound
		}
	}

	// Check if the CLIPath is an executable file
	fileInfo, err := os.Stat(config.CLIPath)
	if err != nil {
		return err
	}
	if (fileInfo.Mode() & 0111) == 0 {
		return errors.New("the CLIPath does not lead to an executable file")
	}

	// TODO: Add more validations here.
	return nil
}
>>>>>>> updated
```

This change adds a check to see if the file at `CLIPath` is executable. If it's not, it returns an error.